requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Foundation Models
transformers>=4.35.0
open-clip-torch>=2.23.0
timm>=0.9.12
segment-anything>=1.0

# Detectron2 (install separately - see README)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
h5py>=3.9.0

# PEFT and Efficient Training
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
loralib>=0.1.1

# Continual Learning Libraries
avalanche-lib>=0.4.0

# Configuration and Experiment Management
omegaconf>=2.3.0
hydra-core>=1.3.0
yacs>=0.1.8
easydict>=1.10

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tensorboard>=2.14.0
wandb>=0.15.0

# Metrics and Evaluation
pycocotools>=2.0.7
scikit-learn>=1.3.0
torchmetrics>=1.2.0

# Model Optimization
onnx>=1.15.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
termcolor>=2.3.0
pyyaml>=6.0.1
fvcore>=0.1.5
portalocker>=2.8.0
cloudpickle>=3.0.0
tabulate>=0.9.0

# Development Tools
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
pytest>=7.4.0
pytest-cov>=4.1.0
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Additional ML Tools
albumentations>=1.3.1
imgaug>=0.4.0
kornia>=0.7.0

# Multi-processing
joblib>=1.3.0

# Type Checking
typing-extensions>=4.8.0
mypy>=1.6.0

# Memory and Performance
psutil>=5.9.0
py3nvml>=0.2.7
gputil>=1.4.0

# Logging
loguru>=0.7.2
